+++ fairmlhealth/__validation.py
@@ -12,7 +12,7 @@ ITER_TYPES = (list, tuple, set, dict, OrderedDict)
 
 
 
-def format_feedback(func):
+def format_errwarn(func):
     """ Wraps a function returning some result with dictionaries for errors and
         warnings, then formats those errors and warnings as grouped warnings.
         Used for reporting functions to skip errors (and warnings) while
@@ -192,10 +192,12 @@ def __validate_length(data, name="array", expected_len=0):
     Raises:
         ValidationError
     """
-    minlen = 4
+    # AIF360's consistency_score defaults to 5 nearest neighbors, thus 5 is
+    #   the minimum acceptable length as long as that dependency exists
+    minlen = 5
     if expected_len < minlen:
         raise ValidationError(f"Cannot measure fewer than {minlen} observations"
-                              + f" (found in {name}")
+                              + f" (found in {name})")
     N = data.shape[0]
     if not N == expected_len:
         raise ValidationError("All data arguments must be of same length."