+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file. Please do f
 
 _“Yesterday I was clever, so I wanted to change the world. Today I am wise, so I am changing myself.”_ - Rumi
 
+## [0.1.12] - 2021-06-17
+### Stratified Report Updates
+- data_report now accepts y as a data frame, enabling stratified analysis across multiple targets.
+- "Overview" column evaluating all features as one unit ("ALL_FEATURES") is now optional via add_overview argument
+
 ## [0.1.11] - 2021-06-17
 ### Validation Updates
 - Validation and preprocessing are now in individual modules to reduce redundancy and improve readability